requirements.txt
Django==2.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
pytz==2017.3
csv
random
nltk
sklearn